A slight shift in direction. Instead of representing types by symbols with namespace-qualified names and putting all symbols in one package, I now map namespaces to packages, and use short names for symbols. This is a much more natural mapping for Lisp. Managing the use of namespaces is now exactly the same as using packages, and tools like SLIME should be able to find symbols in a natural manner, since the scoping maps to what's usually expected in Lisp.
To make this work, type and member symbols had to be distinguishable
to avoid collisions. So member symbol names now start with a period.

Using the alternative reader syntax to do incremental bind-and-import
of types is now much less attractive, except for the fact that it's
parsimonious about creating packages and CLR symbols.

TODO:

The code to re-initialize a system needs rework.  Need to use the term
"binding" consistently in the code to describe the process of setting
up symbols for a type. This is now distinct from "importing", which
now has its conventional Common Lisp meaning.


git-svn-id: https://dmuller.us/sdx/repos/danm/lisp/cl-clr@28 f2fd8bba-85df-0310-9b9a-8e4417b06061