A bit of cleanup, reorg, improvement. Lots of new notes on idiotic things I tried. Actually got the psim test limping along with this version of CL-CLR, although I have a lot of Direct3D stuff to do there yet to get it to display right. This version of CL-CLR uses, essentially, completely dynamic typing via a grody hack around RDNZL. See the first couple of items in TODO.txt for details.
For a while, I thought that direct use of symbols to represent CLR
names was going to be the way to go, and the alternate reader syntax
that I came up with would fall by the wayside. But in practice it's a
pain to use mixed-case symbols in Lisp code. You have to quote them
using vertical bars, like so: |Type|. The reader syntax lets you write
it with just a preceding question mark, ?Type. Surprisingly, this is
easier on the eyes and fingers while programming.

I have a bunch of comments and suggestions for Edi Weitz regarding
RDNZL that I need to write up soon.


git-svn-id: https://dmuller.us/sdx/repos/danm/lisp/cl-clr@30 f2fd8bba-85df-0310-9b9a-8e4417b06061