OK, I still had problems with loading compiled files. This was very educational, although it's not the first time that in my Lisp programming 'career' that I've stumbled over this. It takes a while to wrap one's head around the distinct read, compile, load, and evaluation contexts.
To achieve what I wanted, I needed to take a somewhat different
approach than before, trying to think like a C# compiler. As before,
the reader resolves symbol references, but now it remembers them. The
user has to make a macro call at the end of a file, BIND-CLR-SYMBOLS,
that expands to some forms that will set up all of the symbols that
the reader saw when the compiled file is loaded. This is analogous to
a C# compiling recording binding information about referenced types in
the compiled assembly. Here in Lisp-land, we need just this little bit
of help from the programmer using CL-CLR to make sure that this
happens at the rest time and place.


git-svn-id: https://dmuller.us/sdx/repos/danm/lisp/cl-clr@36 f2fd8bba-85df-0310-9b9a-8e4417b06061