Massive progress. The new substrate, written in C++/CLI, basically works for simple tests. I have high hopes for its efficiency.
Added a tests subdirectory with the basics of a unit test framework.

There's some incomplete cruft in symbols.lisp that was the motivation
for switching focus to a new substrate.

My head hurts from constantly switching between Lisp, C++, C (for the
DLL interface functions), and C# (for a test project on the side where
I investigate details of the runtime's behavior). Oh, and I keep
switching between Emacs and Visual Studio, too. Ow ow ow.


git-svn-id: https://dmuller.us/sdx/repos/cl-clr/trunk@78 f2fd8bba-85df-0310-9b9a-8e4417b06061