Cleaned up the package definition and checked the selection of exported symbols.
Improvemed error checking throughout, both in the Lisp code and in the
substrate. Most of the foreign functions are now able to tell the
caller about exceptions that were thrown (without letting them be
thrown across the Lisp/C boundary, which is important). The conversion
of an exception return to a Lisp condition happens in just one place;
that conversion could still be improved.

Disconnected the reader syntax enablement from manipulation of the
namespace using-list.


git-svn-id: https://dmuller.us/sdx/repos/cl-clr/trunk@82 f2fd8bba-85df-0310-9b9a-8e4417b06061