Got callback mechanisms basically working. Tested only by hand so far.
Noticed problems with handle allocation and deallocation, spent a lot
of time rearranging low-level code and going over it with a
fine-toothed comb. Added some instrumention in the substrate and on
the Lisp side to keep track of handle and CLR-OBJECT counts. Added a
check at the end of the unit tests for lost handles.


git-svn-id: https://dmuller.us/sdx/repos/cl-clr/trunk@88 f2fd8bba-85df-0310-9b9a-8e4417b06061