Fixed the problem with printing CLR-EXCEPTION objects in the debugger.
Added external functions that allow working around the limitation that
symbols-as-types can now only be used with specific CL-CLR functions,
but not in arbitrary places where CLR methods expect types. Not
entirely happy with the results. Rethinking how static symbols and
types are represented in the reader syntax.

Discovered a problem with conversion of negative numbers. Fixed the
logic for this in ffi.lisp. Should write some unit tests for this, but
didn't.

Updated the README.txt a little.


git-svn-id: https://dmuller.us/sdx/repos/cl-clr/trunk@93 f2fd8bba-85df-0310-9b9a-8e4417b06061